api_get_python/requirements.txt
Flask==2.1.0
python-dotenv
psycopg2-binary
